CORE-14335: move KeyOrderBy to libs/crypto/crypto-core so that it can be used in CryptoService
