+++ components/crypto/crypto-service-impl/src/main/kotlin/net/corda/crypto/service/impl/bus/CryptoOpsBusProcessor.kt
@@ -16,7 +16,7 @@ import net.corda.crypto.impl.retrying.CryptoRetryingExecutor
 import net.corda.crypto.impl.toMap
 import net.corda.crypto.impl.toSignatureSpec
 import net.corda.crypto.persistence.SigningKeyInfo
-import net.corda.crypto.service.KeyOrderBy
+import net.corda.crypto.core.KeyOrderBy
 import net.corda.crypto.service.SigningService
 import net.corda.data.crypto.SecureHashes
 import net.corda.data.crypto.ShortHashes
