+++ libs/crypto/crypto-core/src/main/kotlin/net/corda/crypto/core/KeyOrderBy.kt
@@ -1,4 +1,4 @@
-package net.corda.crypto.service
+package net.corda.crypto.core
 
 enum class KeyOrderBy {
     NONE,
